backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pydantic-settings
PyJWT[crypto]>=2.8.0
python-multipart>=0.0.6